images_to_pdf/pyproject.toml
[project]
name = "images-to-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.1.0",
    "streamlit>=1.44.1",
]
